Markers, Backups, TextEdit, and Note Times
Added features:
1) Time markers can be set, to be recalled later when entering a task.
This makes recording accurate project times much easier, rather than
having to remember when a task was started, and manually calculate the
minutes spent on it.
2) Added a backup feature to protect CSV data. The data passed around
in the tasklog.csv has been getting deleted, which is problematic.
Doubly so, because every step of the process manually updates this. So
I added a HARD backup system, which only activates when users call it.
3) Editing Entries now opens the notes in TextEdit! This means that if
the notes need to be edited, they aren’t erased first, and that novel
you wrote last week will remain intact.
4) Notes now include an automatic time range. Preceding notes, there is
a timestamp range like “(11:45AM - 12:22PM)”. This is based on the
entry time, minus the minutes spent. Not too concerned about the exact
accuracy, since this can be edited if something is wrong.